Step 7 (2) : Migrating image info code into slider.js
