+++ src/gallery/slider.js
@@ -1,12 +1,15 @@
 import { msg } from '../i18n'
+import { requestPageMediaInfo } from '../api'
 
 // internal state of the slider component
 let current = 0,
-	dir = ''
+	dir = '',
+	lang,
+	gallery
 
 const clientWidth = window.innerWidth,
 	prefixClassname = 'wp-gallery-fullscreen-slider',
-	renderImageSlider = ( images = [], selectedImage = '', lang, givenDir ) => {
+	renderImageSlider = ( images = [], selectedImage = '', givenLang, givenDir ) => {
         const selectedIndex = images.findIndex( image => image.thumb === selectedImage ), // eslint-disable-line
 			imageListHtml = images.map( ( image ) => `
                 <div class="${prefixClassname}-item">
@@ -16,21 +19,21 @@ const clientWidth = window.innerWidth,
                                 <div class="${prefixClassname}-item-loading-spinner-animation-bounce"></div>
                             </div>
                         </div>
-                        <div class="${prefixClassname}-item-loading-text">${msg( lang, 'gallery-loading-still' )}</div>
+                        <div class="${prefixClassname}-item-loading-text">${msg( givenLang, 'gallery-loading-still' )}</div>
                     </div>
                     <div class="${prefixClassname}-item-loading-error">
-                            <div class="${prefixClassname}-item-loading-error-text">${msg( lang, 'gallery-loading-error' )}</div>
-                            <div class="${prefixClassname}-item-loading-error-refresh">${msg( lang, 'gallery-loading-error-refresh' )}</div>
+                            <div class="${prefixClassname}-item-loading-error-text">${msg( givenLang, 'gallery-loading-error' )}</div>
+                            <div class="${prefixClassname}-item-loading-error-refresh">${msg( givenLang, 'gallery-loading-error-refresh' )}</div>
                         </div>
 					<img src="${image.src}" loading="lazy"/>
-					${ image.caption ? `<div class="${prefixClassname}-item-caption">${image.caption}</div>` : ''}
                 </div>
                 `.trim()
 			).join( '' )
 
-		// set the internal state
 		current = selectedIndex
 		dir = givenDir
+		lang = givenLang
+		gallery = images
 
 		return `
             <div class="${prefixClassname}" style="${dir === 'ltr' ? 'margin-left' : 'margin-right'}:-${selectedIndex * clientWidth}px">
@@ -40,6 +43,47 @@ const clientWidth = window.innerWidth,
             </div>
         `.trim()
 	},
+
+	renderImageInfo = ( mediaInfo, image, lang ) => {
+		const getImageDescription = () => {
+				if ( image.caption ) {
+					return image.caption
+				} else if ( mediaInfo.description ) {
+					return mediaInfo.description
+				} else {
+					return ''
+				}
+			},
+
+			getLicenseInfo = ( license ) => {
+				const licenseTypes = [ 'CC', 'BY', 'SA', 'Fair', 'Public' ]
+				let licenses = ''
+				licenseTypes.forEach( type => {
+					if ( license && license.indexOf( type ) !== -1 ) {
+						licenses += `<div class="${prefixClassname}-item-attribution-info-${type.toLowerCase()}"></div>`
+					}
+				} )
+				return licenses
+			},
+
+			author = mediaInfo.author ? mediaInfo.author : msg( lang, 'gallery-unknown-author' ),
+			link = mediaInfo.filePage
+
+		// @todo consider a wrapper container for all the image info?
+		return `
+			<div class="${prefixClassname}-item-caption">${getImageDescription()}</div>
+			<div class="${prefixClassname}-item-attribution">
+				<div class="${prefixClassname}-item-attribution-info">
+					${getLicenseInfo( mediaInfo.license )}
+					${ author ? `<div class="${prefixClassname}-item-attribution-info-author">${author}</div>` : ''}
+				</div>
+				<div class="${prefixClassname}-item-attribution-more-info">
+					<a href="${link}" class="${prefixClassname}-item-attribution-more-info-link" target="_blank"></a>
+				</div>
+			</div>
+		`.trim()
+	},
+
 	bindImageEvent = ( container, refresh = false ) => {
 		const imageElement = container.querySelector( 'img' ),
 			loading = container.querySelector( `.${prefixClassname}-item-loading` ),
@@ -78,6 +122,7 @@ const clientWidth = window.innerWidth,
 
 		}
 	},
+
 	renderNext = ( offset = 1 ) => {
 		const slider = document.querySelector( `.${prefixClassname}` ),
 			items = slider.querySelectorAll( `.${prefixClassname}-item` ),
@@ -92,13 +137,27 @@ const clientWidth = window.innerWidth,
 			previousButton.style.opacity = current === 0 ? '0.5' : '1'
 
 			bindImageEvent( item )
+
+			// render image attribution element
+			requestPageMediaInfo(
+				lang,
+				gallery[ next ].title,
+				gallery[ next ].fromCommon,
+				mediaInfo => {
+					item.insertAdjacentHTML(
+						'beforeend',
+						renderImageInfo( mediaInfo, gallery[ next ], lang
+						) )
+				} )
 		}
 
 		slider.style[ dir === 'ltr' ? 'marginLeft' : 'marginRight' ] = -clientWidth * current + 'px'
 	},
+
 	renderPrevious = () => {
 		renderNext( -1 )
 	},
+
 	applyGestureEvent = () => {
 		let temp = {
 			screenX: null,
@@ -139,6 +198,7 @@ const clientWidth = window.innerWidth,
 			}
 		} )
 	},
+
 	onShowFn = () => {
 		const sliderContainer = document.querySelector( `.${prefixClassname}` ),
 			items = sliderContainer.querySelectorAll( `.${prefixClassname}-item` ),
