+++ style/gallery.less
@@ -64,6 +64,7 @@
 	&-slider {
 		display: inline-flex;
 		transition: margin-left 0.25s ease-in-out;
+		overflow: hidden;
 
 		&::-webkit-scrollbar {
 			display: none;
@@ -87,7 +88,6 @@
 			}
 
 			&-caption {
-				// @todo the actual position will be updated after image attribution
 				position: absolute;
 				bottom: 35px;
 				width: calc( 100vw - 32px );
@@ -103,6 +103,80 @@
 
 				&::-webkit-scrollbar {
 					display: none;
+				}
+			}
+
+			&-attribution {
+				position: absolute;
+				bottom: 0;
+				width: calc( 100vw - 32px );
+				font-size: 16px;
+				font-weight: normal;
+				font-stretch: normal;
+				font-style: normal;
+				letter-spacing: normal;
+				color: #f8f9fa;
+				border: #f8f9fa;
+				height: 44px;
+				display: flex;
+				justify-content: space-between;
+				margin-bottom: 15px;
+
+				&-info {
+					display: flex;
+					align-items: center;
+
+					&-cc {
+						background-image: inline-svg( '../images/attribution-cc.svg' );
+						&:extend( .license );
+					}
+
+					&-by {
+						background-image: inline-svg( '../images/attribution-author.svg' );
+						&:extend( .license );
+					}
+
+					&-sa {
+						background-image: inline-svg( '../images/attribution-icon-share-a-like.svg' );
+						&:extend( .license );
+					}
+
+					&-fair {
+						background-image: inline-svg( '../images/attribution-license-generic-gnu-free.svg' );
+						&:extend( .license );
+					}
+
+					&-public {
+						background-image: inline-svg( '../images/attribution-license-public-domain.svg' );
+						&:extend( .license );
+					}
+
+					&-author {
+						width: 220px;
+						height: 14px;
+						margin-right: 4px;
+						font-family: 'Lato', 'Linux Libertine', 'Georgia', 'Times', serif;
+						font-size: 12px;
+						color: #fff;
+						overflow: hidden;
+						text-overflow: ellipsis;
+						white-space: nowrap;
+					}
+				}
+
+				&-more-info {
+					height: 44px;
+					width: 44px;
+					display: flex;
+					background-image: inline-svg( '../images/attribution-icon-info.svg' );
+					background-repeat: no-repeat;
+					background-position: center;
+
+					&-link {
+						display: block;
+						height: 44px;
+						width: 44px;
+					}
 				}
 			}
 
@@ -404,73 +478,6 @@
 					-ms-transform: scale( 1 );
 					transform: scale( 1 );
 				}
-			}
-		}
-	}
-
-	&-attribution {
-		border: #f8f9fa;
-		height: 44px;
-		display: flex;
-		justify-content: space-between;
-		margin-bottom: 15px;
-
-		&-info {
-			display: flex;
-			align-items: center;
-			margin-left: 16px;
-
-			&-cc {
-				background-image: inline-svg( '../images/attribution-cc.svg' );
-				&:extend( .license );
-			}
-
-			&-by {
-				background-image: inline-svg( '../images/attribution-author.svg' );
-				&:extend( .license );
-			}
-
-			&-sa {
-				background-image: inline-svg( '../images/attribution-icon-share-a-like.svg' );
-				&:extend( .license );
-			}
-
-			&-fair {
-				background-image: inline-svg( '../images/attribution-license-generic-gnu-free.svg' );
-				&:extend( .license );
-			}
-
-			&-public {
-				background-image: inline-svg( '../images/attribution-license-public-domain.svg' );
-				&:extend( .license );
-			}
-
-			&-author {
-				width: 220px;
-				height: 14px;
-				margin-right: 4px;
-				font-family: 'Lato', 'Linux Libertine', 'Georgia', 'Times', serif;
-				font-size: 12px;
-				color: #fff;
-				overflow: hidden;
-				text-overflow: ellipsis;
-				white-space: nowrap;
-			}
-		}
-
-		&-more-info {
-			margin-right: 16px;
-			height: 44px;
-			width: 44px;
-			display: flex;
-			background-image: inline-svg( '../images/attribution-icon-info.svg' );
-			background-repeat: no-repeat;
-			background-position: center;
-
-			&-link {
-				display: block;
-				height: 44px;
-				width: 44px;
 			}
 		}
 	}
