+++ style/preview.less
@@ -51,12 +51,12 @@
 				}
 			});
 
-			@wordmark-languages-no-margin: ar, bn, fa, he, ps, sd, uz, zh-c;
-			each(@wordmark-languages-no-margin, {
-				&-@{value} {
-					margin: 0 -13px;
-				}
-			});
+			// @wordmark-languages-no-margin: ar, bn, fa, he, ps, sd, uz, zh-c;
+			// each(@wordmark-languages-no-margin, {
+			// 	&-@{value} {
+			// 		margin: 0 -13px;
+			// 	}
+			// });
 		}
 
 		&-closebtn {
